[BENDER-22] correct to uint16_t TwoDRead(const uint16_t* arr, int i, int j, int ny)

+++ bender/Bender.Lib.Native/bender1.cpp
@@ -17,7 +17,7 @@ double TwoDRead(double* arr, int i, int j, int ny)
 	return arr[i * ny + j];
 }
 
-double TwoDRead(const uint16_t* arr, int i, int j, int ny)
+uint16_t TwoDRead(const uint16_t* arr, int i, int j, int ny)
 {
 	return arr[i * ny + j];
 }
